Se realiza correción ejercicio 20 en clase

+++ 02-Ejercicios/Conversion_temperatura.java
@@ -3,15 +3,16 @@ import java.util.Scanner;
 public class Conversion_temperatura{
     public static void main(String []args){
     Scanner entrada = new Scanner(System.in);
-    int Celsius = 0;
-    int Fahrenheit = 0;
+    double Celsius = 0;
+    double Fahrenheit = 0;
 
 
     System.out.println("Ingrese la temperatura en grados Celsius: ");
-    Celsius = entrada.nextInt();
+    Celsius = entrada.nextFloat();
     
     Fahrenheit = ((Celsius * 9/5) + 32);
-     System.out.println("La temperatura en grados Fahrenheit es: "+Fahrenheit);
+    System.out.println("La temperatura en grados Fahrenheit es: "+Fahrenheit);
+
 
 }
